refactor: split into models
